ucm2: MediaTek: mt8390-evk: Add dynamic configuration for HDMI/DP
If mt8390-evk doesn't load the HDMI/DP dtbo, the 'unknown'
available status will be shown on the Ubuntu GNOME.
To solve this issue, add the dynamic configuration to detect
whether the system is loading HDMI/DP dtbo or not.
The HDMI/DP device would be configured only if the related
dtbo has been loaded by the system.

Signed-off-by: Zoran Zhan <zoran.zhan@mediatek.com>